backend: chess/tasks calculate not by flipping

+++ backend/game/gme_chess/tasks/update_calculate_master_with_best_score.py
@@ -7,11 +7,6 @@ from game.models import (
     ChessBoardResultMaster,
     ChessMoveRequestMaster,
 )
-from game.gme_chess.utils import (
-    get_board_from_hash,
-    get_flipped_board,
-    get_hash_from_board,
-)
 from general.gnl_syslog.utils import write_syslog
 from random import randint
 
@@ -24,13 +19,15 @@ def update_calculate_master_with_best_score(calculate_master_id,
     calculate_master = ChessBoardCalculateMaster.objects.get(
         id=calculate_master_id
     )
+    is_upper_side = calculate_master.is_upper_side
 
     def w_debug(message):
         if not is_debug:
             return
         name = 'update_calculate_master_with_best_score'
         level, user = calculate_master.level, calculate_master.created_by
-        message = f'{calculate_master.id}({level}): {message}. [{calculate_master.board}]'
+        message = f'{calculate_master.id}({level})[{is_upper_side}]:' +\
+            f' {message}. [{calculate_master.board}]'
         return write_syslog(name, message, user)
 
     # if i am calculated return~
@@ -48,11 +45,11 @@ def update_calculate_master_with_best_score(calculate_master_id,
         return
 
     # get worst score of children and flip the negative sign as my score
-    children_flipped_scores = [
-        -child_calculate_master.score
+    children_scores = [
+        child_calculate_master.score
         for child_calculate_master in calculate_master.children.all()
     ]
-    best_score = max(children_flipped_scores)
+    best_score = get_best_score(children_scores, is_upper_side)
     w_debug(f'best_score: {best_score}')
 
     # storing result
@@ -64,10 +61,8 @@ def update_calculate_master_with_best_score(calculate_master_id,
     if calculate_master.parent:
         parent_id = calculate_master.parent.id
         w_debug(f'trigger parent: {parent_id}')
-        update_calculate_master_with_best_score.apply_async((
-            parent_id,
-            is_debug,
-        ))
+        update_calculate_master_with_best_score\
+            .apply_async((parent_id, is_debug,))
         return
 
     # no parent. i am root.
@@ -78,21 +73,23 @@ def update_calculate_master_with_best_score(calculate_master_id,
 
     update_request_master(calculate_master.move_request_master, desired_board)
     create_result_master(calculate_master, desired_children)
-    delete_calculate_masters(calculate_master.move_request_master)
+    if not is_debug:
+        delete_calculate_masters(calculate_master.move_request_master)
     w_debug(f'finished')
 
 
 def create_result_master(calculate_master, desired_children):
-    to_boards = [
-        get_flipped_board_hash(desired_child.board)
-        for desired_child in desired_children
-    ]
+    to_boards = [desired_child.board for desired_child in desired_children]
     return ChessBoardResultMaster.objects.create(
         name=calculate_master.board,
         from_board=calculate_master.board,
         to_boards=to_boards,
         created_by=calculate_master.created_by
     )
+
+
+def get_best_score(scores, is_upper_side):
+    return max(scores) if is_upper_side else min(scores)
 
 
 def delete_calculate_masters(request_master):
@@ -107,13 +104,7 @@ def get_desired_children(children_masters, best_score):
 def get_desired_board(desire_children, is_random):
     desire_count = desire_children.count()
     desire_index = randint(0, desire_count - 1) if is_random else 0
-    return get_flipped_board_hash(desire_children[desire_index].board)
-
-
-def get_flipped_board_hash(board_hash):
-    board = get_board_from_hash(board_hash)
-    flipped_board = get_flipped_board(board)
-    return get_hash_from_board(flipped_board)
+    return desire_children[desire_index].board
 
 
 def update_request_master(request_master, desired_board):
